Add multiple k values support

+++ bin_x/cli/clustering.py
@@ -4,67 +4,51 @@ from pathlib import Path
 from typing import Optional
 
 import click
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import seaborn as sns
-from numpy.lib.format import open_memmap
+from numpy.lib.format import open_memmap  # noqa
 
-from bin_x.cli.utils import handle_error, reduce_dimensions_to_2d
 from bin_x.core.clustering.algorithm import fit_cluster
-from bin_x.core.clustering.distance_matrix import create_distance_matrix
-from bin_x.core.config import USER_CONFIG
-
-
-def _visualize_final_result(
-    num_clusters: int, df_features: pd.DataFrame, df_dist_bin: pd.DataFrame, split_bin_png: Path, bin_counts_png: Path
-):
-    df_kmer_bins = df_dist_bin.merge(df_features, left_on="CONTIG_NAME", right_on="PARENT_NAME")
-    df_kmer_counts = df_kmer_bins.drop(["CONTIG_NAME_x", "CONTIG_NAME_y", "PARENT_NAME", "CLUSTER", "BIN"], axis=1)
-    df_2d = reduce_dimensions_to_2d(df_kmer_counts)
-
-    plt.figure(figsize=(15, 10))
-    plt.title("Convex hull Result (Split)")
-    for i in range(-1, num_clusters):
-        sns.scatterplot(x="X", y="Y", data=df_2d[df_kmer_bins.BIN == i], label=f"Cluster {i}", alpha=0.1)
-    plt.savefig(split_bin_png)
-    plt.close()
-
-    plt.figure(figsize=(15, 10))
-    plt.title("Bin Counts")
-    sns.histplot(df_kmer_bins, x="BIN", discrete=True)
-    plt.savefig(bin_counts_png)
-    plt.close()
+from bin_x.core.clustering.distance_matrix import (
+    create_distance_matrix,
+    create_in_mem_distance_matrix,
+)
+from bin_x.core.clustering.dump_bins import dump_bins
 
 
 def perform_clustering(
+    contig_fasta: Path,
     features_csv: Path,
     operating_dir: Path,
     num_neighbors: int = 15,
     max_iterations: int = 10,
     metric: str = "convex",
     qp_solver: str = "quadprog",
+    in_mem_dist_matrix: bool = True,
     distance_matrix_cache: Optional[Path] = None,
 ) -> Path:
     """
     Perform binning and output the binning result.
 
+    :param contig_fasta: Contig file used for feature generation.
     :param features_csv: CSV containing the feature vectors and initial bins.
     :param operating_dir: Directory to write temp files to.
     :param num_neighbors: Number of neighbors to consider for polytope.
     :param max_iterations: Number of maximum iterations to perform.
     :param metric: Polytope distance matrix (convex/affine)
     :param qp_solver: Quadratic programming problem solver. (quadprog/cvxopt)
+    :param in_mem_dist_matrix: Whether to use in memory distance matrix.
     :param distance_matrix_cache: Previously computed distance matrix.
+                                    in_mem_dist_matrix must be false if this is provided.
     :return: Path of the binning result dataset.
     """
-    dist_bin_csv = operating_dir / "bin.csv"
-    split_bin_png = operating_dir / "split_bin.png"
-    bin_counts_png = operating_dir / "bin_counts.png"
+    dist_bin_csv = operating_dir / "binning-assignment.csv"
+    bin_dump_dir = operating_dir / "bins"
     operating_dir.mkdir(parents=True, exist_ok=True)
+    bin_dump_dir.mkdir(parents=True, exist_ok=True)
 
     # 01. Read feature CSV
-    click.secho("01. Reading feature CSV...", bold=True)
+    click.secho(">> Reading feature CSV...", fg="green", bold=True)
     df_features = pd.read_csv(features_csv)
 
     num_clusters = df_features.CLUSTER.max() + 1
@@ -74,15 +58,20 @@ def perform_clustering(
 
     # 02. Create a distance matrix
     distance_matrix_filename = distance_matrix_cache
-    if distance_matrix_filename is None:
-        click.secho(f"02. Creating a distance matrix of {num_samples}x{num_samples} shape...", bold=True)
-        distance_matrix_filename = create_distance_matrix(samples, operating_dir)
+    if in_mem_dist_matrix:
+        assert distance_matrix_filename is None, "Distance matrix cache cannot be used in in-memory mode"
+        click.secho(f">> Creating the distance matrix({num_samples}x{num_samples}) in-memory...", fg="green", bold=True)
+        distance_matrix = create_in_mem_distance_matrix(samples)
     else:
-        click.secho(f"02. Reusing a distance matrix at {distance_matrix_filename}...", bold=True)
-    distance_matrix = open_memmap(filename=distance_matrix_filename, mode="r", shape=(num_samples, num_samples))
+        if distance_matrix_filename is None:
+            click.secho(f">> Creating a distance matrix of {num_samples}x{num_samples} shape...", fg="green", bold=True)
+            distance_matrix_filename = create_distance_matrix(samples, operating_dir)
+        else:
+            click.secho(f">> Reusing distance matrix at {distance_matrix_filename}...", fg="green", bold=True)
+        distance_matrix = open_memmap(filename=distance_matrix_filename, mode="r", shape=(num_samples, num_samples))
 
     # 03. Perform binning using specified solver and metric
-    click.secho(f"03. Performing binning using {qp_solver} solver", bold=True)
+    click.secho(f">> Performing binning using {qp_solver} solver", fg="green", bold=True)
     convex_labels = fit_cluster(
         samples=samples,
         num_clusters=num_clusters,
@@ -95,13 +84,14 @@ def perform_clustering(
     )
 
     # Delete the distance matrix file (dont delete if using a cache)
-    if distance_matrix_cache is None:
+    if not in_mem_dist_matrix and distance_matrix_cache is None:
+        assert distance_matrix_filename is not None
         os.remove(distance_matrix_filename)
     if np.any(convex_labels < 0):
         raise ValueError("There were some un-clustered points left... Aborting.")
 
     # 04. Assigning bins with majority voting (If there were more than one voting column)
-    click.secho("04. Assigning bins", bold=True)
+    click.secho(">> Assigning bins", fg="green", bold=True)
     df_samples: pd.DataFrame = df_features.drop("CLUSTER", axis=1)
     df_bin_column: pd.DataFrame = pd.DataFrame({"BIN": convex_labels})
 
@@ -110,16 +100,18 @@ def perform_clustering(
     df_dist_bin: pd.DataFrame = parent_groups.BIN.apply(lambda x: np.bincount(x).argmax()).reset_index()
     df_dist_bin.rename(columns={"PARENT_NAME": "CONTIG_NAME"}, inplace=True)
     df_dist_bin.to_csv(dist_bin_csv, index=False)  # noqa
-    click.secho(f"Dumped binning assignment CSV at {dist_bin_csv}", fg="green", bold=True)
+    click.secho(f"Dumped binning assignment CSV at {dist_bin_csv}", bold=True)
 
-    # 05. Visualize
-    click.secho("08. Drawing plots...", bold=True)
-    _visualize_final_result(num_clusters, df_features, df_dist_bin, split_bin_png, bin_counts_png)
+    # 05. Creating bin files with contigs
+    click.secho(">> Writing binned FASTA files", fg="green", bold=True)
+    dump_bins(df_dist_bin, contig_fasta, bin_dump_dir)
+    click.secho(f"Dumped binned fasta to {bin_dump_dir}", bold=True)
 
     return dist_bin_csv
 
 
 def run_perform_clustering(
+    contig_fasta: Path,
     features_csv: Path,
     operating_dir: Path,
     parameters: SectionProxy,
@@ -128,7 +120,7 @@ def run_perform_clustering(
     """
     Perform binning and output the binning result.
 
-
+    :param contig_fasta: Contig file used for feature generation.
     :param features_csv: CSV containing the feature vectors and initial bins.
     :param operating_dir: Directory to write temp files to.
     :param parameters: Parameters INI section.
@@ -137,29 +129,13 @@ def run_perform_clustering(
     """
 
     return perform_clustering(
+        contig_fasta=contig_fasta,
         features_csv=features_csv,
         operating_dir=operating_dir,
         num_neighbors=int(parameters["AlgoNumNeighbors"]),
         max_iterations=int(parameters["AlgoMaxIterations"]),
         metric=parameters["AlgoDistanceMetric"],
         qp_solver=parameters["AlgoQpSolver"],
+        in_mem_dist_matrix=parameters.getboolean("InMemDistMatrix"),
         distance_matrix_cache=distance_matrix_cache,
     )
-
-
-@click.command()
-@click.option("--config", prompt="Configuration file", help="The INI File to use for tool configuration.", type=Path)
-@click.option("--features", prompt="Features file", help="The features CSV file with initial clustering.", type=Path)
-@click.option("--out", prompt="Output Directory", help="The output directory for the tool.", type=Path)
-def main(config: Path, features: Path, out: Path):
-    try:
-        np.random.seed(0)
-        USER_CONFIG.read(config)
-        parameters = USER_CONFIG["PARAMETERS"]
-        run_perform_clustering(features, out, parameters)
-    except Exception as e:
-        handle_error(e)
-
-
-if __name__ == "__main__":
-    main()
